+++ .gitignore
@@ -38,10 +38,13 @@ n_app/config/master.key
 n_app/app/assets/builds/*
 !n_app/app/assets/builds/.keep
 
-n_app/.gitignore
-/node_modules
-/config/local_env.yml
-n_app/node_modules
-n_app/config/local_env.yml
-
+# Ignore seeds as it is used for development and testing purposes only
 n_app/db/seeds.rb
+
+# Ignore uploads made
+n_app/public/uploads
+
+/node_modules
+n_app/node_modules
+
+n_app/.gitignore
