+++ n_app/app/controllers/members/invitations_controller.rb
@@ -2,7 +2,7 @@ class Members::InvitationsController < Devise::InvitationsController
   before_action :configure_permitted_parameters
 
   def after_invite_path_for(resource)
-    member_editor_path
+    site_admins_members_manager_index_path
   end
 
   def update
