--- n_app/app/controllers/site_admins/carousel_images.rb
@@ -1,2 +1,0 @@
-# frozen_string_literal: true
-
