--- n_app/app/controllers/site_admins/event_histories.rb
@@ -1,2 +1,0 @@
-# frozen_string_literal: true
-
