+++ n_app/app/models/event_history.rb
@@ -1,6 +1,4 @@
 class EventHistory < ApplicationRecord
-    validates :event_title, presence: true
-    validates :event_text, presence: true
-    validates :date_of_event, presence: true
+    validates :event_title, :event_text, :date_of_event, presence: true
     default_scope { order(date_of_event: :desc) }
 end
