--- n_app/app/views/site_admins/admin_index.erb
@@ -1,30 +1,0 @@
-<h1>Welcome Admin</h1>
-<br />
-<div class="container text-center">
-  <div class="row row-cols-2 row-cols-lg-4 g-2 g-lg-5">
-    <div class="col">
-        <%= link_to 'Member Editor', editor_site_admins_members_path, class:"p-3 btn btn-light d-block" %>
-    </div>
-    <div class="col">
-        <%= link_to 'Carousel Editor', editor_site_admins_carousel_images_path, class:"p-3 btn btn-light d-block" %>
-    </div>
-    <div class="col">
-        <%= link_to 'Event History Editor', editor_site_admins_event_histories_path, class:"p-3 btn btn-light d-block" %>
-    </div>
-    <div class="col">
-        <%= link_to 'Member Statistics', statistics_site_admins_members_path, class:"p-3 btn btn-light d-block" %>
-    </div>
-    <div class="col">
-        <%= link_to 'Placeholder', site_admins_admin_index_path, class:"p-3 btn btn-light d-block" %>
-    </div>
-    <div class="col">
-        <%= link_to 'Placeholder', site_admins_admin_index_path, class:"p-3 btn btn-light d-block" %>
-    </div>
-    <div class="col">
-        <%= link_to 'Placeholder', site_admins_admin_index_path, class:"p-3 btn btn-light d-block" %>
-    </div>
-    <div class="col">
-        <%= link_to 'Placeholder', site_admins_admin_index_path, class:"p-3 btn btn-light d-block" %>
-    </div>
-  </div>
-</div>
