+++ n_app/app/views/site_admins/base/index.html.erb
@@ -1,0 +1,30 @@
+<h1>Welcome Admin</h1>
+<br />
+<div class="container text-center">
+    <div class="row row-cols-2 row-cols-lg-4 g-2 g-lg-5">
+        <div class="col">
+            <%= link_to 'Member Editor', site_admins_members_manager_index_path, class:"p-3 btn btn-light d-block" %>
+        </div>
+        <div class="col">
+              <%= link_to 'Carousel Editor', site_admins_carousel_images_path, class:"p-3 btn btn-light d-block" %>
+        </div>
+        <div class="col">
+            <%= link_to 'Event History Editor', site_admins_event_histories_path, class:"p-3 btn btn-light d-block" %>
+        </div>
+        <div class="col">
+            <%= link_to 'Member Statistics', statistics_site_admins_members_manager_index_path, class:"p-3 btn btn-light d-block" %>
+        </div>
+        <div class="col">
+            <%= link_to 'Placeholder', site_admins_index_path, class:"p-3 btn btn-light d-block" %>
+        </div>
+        <div class="col">
+            <%= link_to 'Placeholder', site_admins_index_path, class:"p-3 btn btn-light d-block" %>
+        </div>
+        <div class="col">
+            <%= link_to 'Placeholder', site_admins_index_path, class:"p-3 btn btn-light d-block" %>
+        </div>
+        <div class="col">
+            <%= link_to 'Placeholder', site_admins_index_path, class:"p-3 btn btn-light d-block" %>
+        </div>
+    </div>
+</div>
