--- n_app/app/views/site_admins/member_editor.html.erb
@@ -1,132 +1,0 @@
-<div class="table-responsive">
-  <table class="table table-bordered table-light table-hover table-sm table align-middle caption-top">
-    <caption><i class="bi bi-person-lines-fill"></i> List of Users</caption>
-    <thead class="table-secondary">
-      <tr>
-        <th scope="col">Student ID</th>
-        <th scope="col">Member Name</th>
-        <th scope="col">Year</th>
-        <th scope="col">Department</th>
-        <% if current_member.member_role == "admin" %>
-          <th scope="col">Access Level</th>
-          <th scope="col">Member Status</th>
-        <% end %>
-      </tr>
-    </thead>
-    <tbody>
-      <% @members.each do |member| %>
-        <tr>
-          <th scope="row"><%= member.student_id %></th>
-          <td>
-            <%= link_to member.name, member_path(member.id), class: 'link-offset-2 link-underline link-underline-opacity-0' %>
-          </td>
-          <td><%= member.grade %></td>
-          <td><%= member.department %></td>
-          <% if current_member.member_role == "admin" %>
-            <% if not member.invitation_accepted_at %>
-              <td class="table-light"><div class="full-width-container">
-                <%= button_to "<i class=\"bi bi-envelope-at\"></i> Resend Invitation".html_safe, resend_invitation_path(member.id), method: :get, class:"btn btn-sm full-width-button" %>
-              </div></td>
-            <% elsif member.member_role == "user" %>
-              <td class="table-primary"><div class="full-width-container">
-                <%= button_to "<i class=\"bi bi-check-square\"></i> Grant Mod Privileges".html_safe, grant_mod_status_path(member.id), method: :patch, class:"btn btn-sm full-width-button" %>
-              </div>
-            <% elsif member.member_role == "mod" %>
-              <td class="table-success"><div class="full-width-container">
-                <%= button_to "<i class=\"bi bi-person-check\"></i> Grant Admin Privileges".html_safe, grant_admin_status_path(member.id), method: :patch, class:"btn btn-sm full-width-button" %>
-              </div>
-            <% else %>
-              <td class="table-info"><div class="full-width-container">
-                <button type="button" class="btn btn-sm full-width-button">
-                  <i class="bi bi-exclamation-square"></i> Admin
-                </button>
-              </div>
-            <% end %>
-            <% if member.member_role == "user" %>
-              <td class="table-danger"><div class="full-width-container">
-                <%= link_to destroy_member_path(member.id), data: { turbo_method: :delete, turbo_confirm: "<<<REMOVE MEMBERSHIP>>>\nPROCEED?" }, class:"btn btn-sm full-width-button" do %>
-                  <i class="bi bi-person-x"></i> Remove Membership
-                <% end %>
-              </div></td>
-            <% elsif member.member_role == "mod" %>
-              <td class="table-warning"><div class="full-width-container">
-                <%= button_to "<i class='bi bi-x-square'></i> Revoke Mod Privileges".html_safe, revoke_mod_status_path(member.id), method: :patch, class:"btn btn-sm full-width-button" %>
-              </div>
-            <% elsif member.member_role == "admin" && member != current_member %>
-              <td class="table-warning"><div class="full-width-container">
-                <%= button_to "<i class='bi bi-x-square'></i> Revoke Admin Privileges".html_safe, revoke_admin_status_path(member.id), method: :patch, class:"btn btn-sm full-width-button" %>
-              </div>
-            <% else %>
-              <td class="table-info"><div class="full-width-container">
-                <button type="button" class="btn btn-sm full-width-button">
-                  <i class="bi bi-exclamation-square"></i> Cannot Demote Yourself
-                </button>
-              </div>
-            <% end %>
-          <% end %>
-        </tr>
-      <% end %>
-    </tbody>
-  </table>
-</div>
-<br />
-<div class="card">
-  <div class="card-header">
-    Administrative Functions
-  </div>
-  <ul class="list-group list-group-horizontal d-flex flex-wrap">
-    <li class="flex-fill custom-list-item">
-      <%= button_to "<i class=\"bi bi-arrow-up-square\"></i> INCREMENT ALL GRADES BY 1 YEAR".html_safe, { :controller => "site_admins", :action => "increment_grade"}, data: { turbo_method: :delete, turbo_confirm: "<<<INCREMENT ALL GRADES BY 1 YEAR>>>\nPROCEED?" }, class: "list-group-item list-group-item-action list-group-item-primary w-100 text-center" %>
-    </li>
-    <li class="flex-fill custom-list-item">
-      <%= button_to "<i class=\"bi bi-arrow-down-square\"></i> DECREMENT ALL GRADES BY 1 YEAR".html_safe, { :controller => "site_admins", :action => "decrement_grade"}, data: { turbo_method: :delete, turbo_confirm: "<<<DECREMENT ALL GRADES BY 1 YEAR>>>\nPROCEED?" }, class: "list-group-item list-group-item-action list-group-item-warning w-100 text-center" %>
-    </li>
-    <% if current_member.member_role == "admin" %>
-      <li class="flex-fill custom-list-item">
-        <%= link_to new_member_invitation_path, class: "list-group-item list-group-item-action list-group-item-info w-100 text-center" do %>
-          <i class="bi bi-envelope-at"></i>
-          INVITE NEW MEMBER BY GMAIL
-        <% end %>
-      </li>
-    <% end %>
-    <li class="flex-fill custom-list-item">
-      <%= link_to member_editor_path(format: 'xlsx'), class: "list-group-item list-group-item-action list-group-item-success w-100 text-center" do %>
-        <i class="bi bi-file-earmark-excel"></i>
-        EXPORT AS MICROSOFT EXCEL SPREADSHEET
-      <% end %>
-    </li>
-  </ul>
-</div>
-<br />
-<%= link_to admin_index_site_admins_path, class:"link-offset-2 link-underline link-underline-opacity-0" do %>
-  <i class="bi bi-arrow-left-square"></i>
-  Back To Admin Index
-<% end %>
-
-<style>
-  .custom-list-item {
-    list-style-type: none;
-    padding: 0;
-    margin: 0;
-  }
-
-  .table td {
-    padding: 0;
-    margin: 0;
-  }
-
-  .full-width-container {
-    width: 100%; /* Container takes full width of cell */
-    display: table; /* Ensures the container respects table cell layout */
-    height: 100%; /* Container takes full height of cell */
-  }
-
-  .full-width-button {
-    width: 100%; /* Button takes full width of container */
-    display: table-cell; /* Makes the button respect the table layout */
-    text-align: center; /* Centers text within the button */
-    vertical-align: middle; /* Aligns the button text vertically in the middle */
-    box-sizing: border-box; /* Ensures padding and borders are included in width */
-    margin: 0; /* Removes any default margins */
-  }
-</style>
