--- n_app/app/views/site_admins/member_statistics.html.erb
@@ -1,109 +1,0 @@
-<div class="table-responsive">
-  <table class="table caption-top">
-    <caption><i class="bi bi-clipboard-data-fill"></i> Statistics on Member Grades</caption>
-    <thead>
-      <tr>
-        <th scope="col">Year</th>
-        <th scope="col">Member Count</th>
-        <th scope="col">Percentage</th>
-      </tr>
-    </thead>
-    <tbody>
-      <% @grades.each do |grade| %>
-        <tr>
-          <th scope="row"><%= grade %></th>
-          <td><%= @count[grade].to_i %></td>
-          <% if @memberCount == 0 %>
-            <td>0%</td>
-          <% else %>
-            <td><%= (@count[grade].to_i * 100 / @memberCount.to_i).to_s + "%" %></td>
-          <% end %>
-        </tr>
-      <% end %>
-    </tbody>
-    <tfoot>
-      <tr>
-        <th scope="row">Total</th>
-        <td><%= @memberCount %></td>
-        <td>100%</td>
-      </tr>
-    </tfoot>
-  </table>
-</div>
-<br />
-<div class="table-responsive">
-  <table class="table caption-top">
-    <caption><i class="bi bi-clipboard-data-fill"></i> Statistics on Member Departments</caption>
-    <thead>
-      <tr>
-        <th style="width: 5%;" scope="col">Year</th>
-        <th scope="col">Department</th>
-        <th scope="col">Share</th>
-      </tr>
-    </thead>
-    <tbody>
-      <% @grades.each do |grade| %>
-        <tr>
-          <th scope="row"><%= grade %></th>
-          <td>
-            <ul class="list-group list-group-horizontal">
-              <li class="list-group-item" style="flex: 1;">IT</li>
-              <li class="list-group-item flex-fill" style="flex: 1;"><%= @departmentCount[[grade, "情報工学科"]].to_i %></li>
-            </ul>
-            <ul class="list-group list-group-horizontal">
-              <li class="list-group-item" style="flex: 1;">DE</li>
-              <li class="list-group-item flex-fill" style="flex: 1;"><%= @departmentCount[[grade, "デジタルエンタテインメント学科"]].to_i %></li>
-            </ul>
-          </td>
-          <td>
-            <ul class="list-group">
-              <% if @count[grade].to_i == 0 %>
-                <li class="list-group-item">0%</li>
-                <li class="list-group-item">0%</li>
-              <% else %>
-                <li class="list-group-item"><%= (@departmentCount[[grade, "情報工学科"]].to_i * 100 / @count[grade].to_i).to_s + "%" %></li>
-                <li class="list-group-item"><%= (@departmentCount[[grade, "デジタルエンタテインメント学科"]].to_i * 100 / @count[grade].to_i).to_s + "%" %></li>
-              <% end %>
-            </ul>
-          </td>
-        </tr>
-      <% end %>
-      <tr>
-        <th scope="row">Combined</th>
-        <td>
-          <ul class="list-group list-group-horizontal">
-            <li class="list-group-item" style="flex: 1;">IT</li>
-            <li class="list-group-item flex-fill" style="flex: 1;"><%= @totalDepartmentCount["情報工学科"].to_i %></li>
-          </ul>
-          <ul class="list-group list-group-horizontal">
-            <li class="list-group-item" style="flex: 1;">DE</li>
-            <li class="list-group-item flex-fill" style="flex: 1;"><%= @totalDepartmentCount["デジタルエンタテインメント学科"].to_i %></li>
-          </ul>
-        </td>
-        <td>
-          <ul class="list-group">
-            <% if @memberCount.to_i == 0 %>
-              <li class="list-group-item">0%</li>
-              <li class="list-group-item">0%</li>
-            <% else %>
-              <li class="list-group-item"><%= (@totalDepartmentCount["情報工学科"].to_i * 100 / @memberCount.to_i).to_s + "%" %></li>
-              <li class="list-group-item"><%= (@totalDepartmentCount["デジタルエンタテインメント学科"].to_i * 100 / @memberCount.to_i).to_s + "%" %></li>
-            <% end %>
-          </ul>
-        </td>
-      </tr>
-    </tbody>
-    <tfoot>
-      <tr>
-        <th scope="row">Total</th>
-        <td><%= @memberCount %></td>
-        <td>100%</td>
-      </tr>
-    </tfoot>
-  </table>
-</div>
-<br />
-<%= link_to admin_index_site_admins_path, class:"link-offset-2 link-underline link-underline-opacity-0" do %>
-  <i class="bi bi-arrow-left-square"></i>
-  Back To Admin Index
-<% end %>
